Fix performance issues with many offline sessions
Fixes: #13340

Cora note: there was one conflict during this merge, in
PersistentClientSessionEntity.java. It was resolved by accepting all
incoming changes.

The loss in the resolution was of query `findClientSessionsOrderedById`,
which was replaced by `findClientSessionsOrderedByIdInterval` and
`findClientSessionsOrderedByIdExact` in
PersistentClientSessionEntity.java. I have confirmed that the old query
is no longer referenced.